requirements.txt
openai
opencv-python
numpy
pillow
python-dotenv
playwright
magentic-one[all]